--- client/src/components/Nav-bar/Nav-bar.css
@@ -1,4 +1,0 @@
-.fullbar {
-  background-color: maroon;
-  color: white;
-}
