+++ client/src/components/index.js
@@ -2,7 +2,7 @@ import Content from "./content";
 import Footer from "./footer";
 import Hero from "./hero";
 import Loading from "./loading";
-import NavBar from "./Navbar";
+import NavBar from "./Navbar/Navbar";
 import Highlight from "./highlight";
 import PrivateRoute from "./private-route";
